fix: fix module bugs after the latest refactor.
