--- leaf-stream/src/modules/wasm.rs
@@ -1,399 +1,0 @@
-use std::{
-    iter,
-    mem::size_of,
-    pin::Pin,
-    sync::{Arc, LazyLock},
-};
-
-use anyhow::Context;
-use blake3::Hash;
-use leaf_stream_types::{
-    EventRequest, Inbound, IncomingEvent, ModuleInit, Outbound, Process, SqlQuery, SqlRow, SqlRows,
-    SqlValue,
-};
-use libsql::Connection;
-use parity_scale_codec::{Decode, Encode};
-use wasmtime::{Config, Engine, FuncType, Linker, Module, Store, Val, ValType};
-
-use crate::LeafModule;
-
-#[derive(Debug)]
-pub struct LeafWasmModule {
-    id: Hash,
-    module: wasmtime::Module,
-    linker: Arc<wasmtime::Linker<libsql::Connection>>,
-}
-
-#[derive(Debug)]
-struct InvalidFunctionTypeCtx(&'static str);
-impl std::fmt::Display for InvalidFunctionTypeCtx {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Invalid type for {} function", self.0)
-    }
-}
-#[derive(Debug)]
-struct ErrorCallingFunctionCtx(&'static str);
-impl std::fmt::Display for ErrorCallingFunctionCtx {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Error calling {} function", self.0)
-    }
-}
-
-#[derive(thiserror::Error, Debug)]
-pub enum ModuleError {
-    #[error("Error with WASM module interface: {0}")]
-    InterfaceError(#[from] WasmInterfaceError),
-    #[error("WASM error: {0}")]
-    WasmError(#[from] wasmtime::Error),
-}
-
-#[derive(thiserror::Error, Debug)]
-pub enum WasmInterfaceError {
-    #[error("WASM module does not export the malloc function.")]
-    InvalidOrMissingMallocExport,
-    #[error("WASM module does not export memory.")]
-    MissingMemoryExport,
-    #[error("The process_event function is either not exported or has an invalid type.")]
-    InvalidOrMissingProcessEventExport,
-}
-
-impl LeafWasmModule {
-    pub fn new(wasm: &[u8]) -> Result<Self, ModuleError> {
-        let id = blake3::hash(wasm);
-        let module = wasmtime::Module::new(&ENGINE, wasm)?;
-        let mut linker = wasmtime::Linker::<libsql::Connection>::new(&ENGINE);
-        linker.func_wrap_async(
-            "host",
-            "query",
-            |mut caller, (query_ptr, query_len, out_ptr): (i32, i32, i32)| {
-                let result = async move {
-                    let memory = caller
-                        .get_export("memory")
-                        .and_then(|x| x.into_memory())
-                        .ok_or_else(|| anyhow::format_err!("Could not get WASM memory export"))?;
-                    let Some(malloc) = caller.get_export("malloc").and_then(|x| x.into_func())
-                    else {
-                        anyhow::bail!("WASM module does not export `malloc` function");
-                    };
-                    let conn = caller.data();
-
-                    let mut query_bytes = Vec::from_iter(iter::repeat_n(0u8, query_len as usize));
-                    memory.read(&caller, query_ptr as usize, &mut query_bytes)?;
-                    let query = SqlQuery::decode(&mut &query_bytes[..])?;
-
-                    // TODO: provide a way to execute batch queries.
-                    let mut rows = conn
-                        .query(
-                            &query.sql,
-                            query
-                                .params
-                                .into_iter()
-                                .map(|(key, value)| (key, to_libsql(value)))
-                                .collect::<Vec<_>>(),
-                        )
-                        .await?;
-
-                    let col_count = rows.column_count();
-                    let mut result = SqlRows {
-                        column_names: (0..col_count)
-                            .map(|i| rows.column_name(i).unwrap_or_default().to_owned())
-                            .collect(),
-                        ..Default::default()
-                    };
-                    while let Some(row) = rows.next().await? {
-                        result.rows.push(SqlRow {
-                            values: (0..col_count)
-                                .map(|i| Ok::<_, libsql::Error>(from_libsql(row.get_value(i)?)))
-                                .collect::<Result<Vec<_>, _>>()?,
-                        });
-                    }
-
-                    // Allocate the query result and write it to wasm memory
-                    let result_bytes = result.encode();
-                    let result_len = result_bytes.len() as i32;
-                    let mut results = [Val::null_any_ref()];
-                    malloc
-                        .call_async(&mut caller, &[result_len.into(), 1.into()], &mut results)
-                        .await?;
-                    let Some(result_ptr) = results[0].i32() else {
-                        anyhow::bail!("Invalid return type from malloc");
-                    };
-                    memory.write(&mut caller, result_ptr as usize, &result_bytes)?;
-
-                    // Write the pointer to the result to the out pointer
-                    let ptr_bytes: Vec<u8> = i32::to_le_bytes(result_ptr)
-                        .into_iter()
-                        .chain(i32::to_le_bytes(result_len).into_iter())
-                        .collect();
-                    memory.write(&mut caller, out_ptr as usize, &ptr_bytes)?;
-
-                    Ok(())
-                };
-
-                Box::new(async move { result.await.inspect_err(|e| tracing::error!("{e}")) })
-            },
-        )?;
-
-        Self::validate_module_interface(&module)?;
-
-        Ok(Self {
-            id,
-            module,
-            linker: Arc::new(linker),
-        })
-    }
-
-    fn validate_module_interface(w_module: &wasmtime::Module) -> Result<(), WasmInterfaceError> {
-        if !w_module
-            .exports()
-            .any(|x| x.name() == "memory" && x.ty().memory().is_some())
-        {
-            return Err(WasmInterfaceError::MissingMemoryExport);
-        }
-        if let Some(x) = w_module.get_export("malloc")
-            && let Some(f) = x.func()
-            && f.matches(&FuncType::new(
-                &ENGINE,
-                [ValType::I32, ValType::I32],
-                [ValType::I32],
-            ))
-        {
-        } else {
-            return Err(WasmInterfaceError::InvalidOrMissingMallocExport);
-        };
-        if let Some(x) = w_module.get_export("filter_inbound")
-            && let Some(f) = x.func()
-            && f.matches(&FuncType::new(
-                &ENGINE,
-                [ValType::I32, ValType::I32, ValType::I32],
-                [],
-            ))
-        {
-        } else {
-            return Err(WasmInterfaceError::InvalidOrMissingProcessEventExport);
-        };
-        if let Some(x) = w_module.get_export("filter_outbound")
-            && let Some(f) = x.func()
-            && f.matches(&FuncType::new(
-                &ENGINE,
-                [ValType::I32, ValType::I32, ValType::I32],
-                [],
-            ))
-        {
-        } else {
-            return Err(WasmInterfaceError::InvalidOrMissingProcessEventExport);
-        };
-        if let Some(x) = w_module.get_export("process_event")
-            && let Some(f) = x.func()
-            && f.matches(&FuncType::new(
-                &ENGINE,
-                [ValType::I32, ValType::I32, ValType::I32],
-                [],
-            ))
-        {
-        } else {
-            return Err(WasmInterfaceError::InvalidOrMissingProcessEventExport);
-        };
-
-        Ok(())
-    }
-
-    async fn call_wasm_func<I: Encode, O: Decode>(
-        func_name: &'static str,
-        module: Module,
-        linker: Arc<Linker<Connection>>,
-        db: libsql::Connection,
-        input: I,
-    ) -> anyhow::Result<O> {
-        let mut store = Store::new(&ENGINE, db);
-        let instance = linker.instantiate_async(&mut store, &module).await?;
-        let Some(memory) = instance.get_memory(&mut store, "memory") else {
-            anyhow::bail!("Could not load wasm memory.");
-        };
-        let malloc = instance
-            .get_typed_func::<(i32, i32), i32>(&mut store, "malloc")
-            .context(InvalidFunctionTypeCtx("malloc"))?;
-        let func = instance
-            .get_typed_func::<(i32, i32, i32), ()>(&mut store, func_name)
-            .context(InvalidFunctionTypeCtx(func_name))?;
-
-        // Allocate the input in the WASM module
-        let input_bytes = input.encode();
-        let input_len = input_bytes.len() as i32;
-        let input_ptr = malloc.call_async(&mut store, (input_len, 1)).await?;
-        memory.write(&mut store, input_ptr as usize, &input_bytes)?;
-
-        // Allocate space for the return value
-        let output_ptr = malloc
-            .call_async(
-                &mut store,
-                (size_of::<i32>() as i32 * 2, align_of::<i32>() as i32),
-            )
-            .await?;
-
-        // Run the named function from the WASM module
-        func.call_async(&mut store, (input_ptr, input_len, output_ptr))
-            .await
-            .context(ErrorCallingFunctionCtx(func_name))?;
-
-        let mut output_ptr_bytes = [0u8; size_of::<i32>()];
-        memory.read(&store, output_ptr as usize, &mut output_ptr_bytes)?;
-        let mut output_len_bytes = [0u8; size_of::<i32>()];
-        memory.read(
-            &store,
-            output_ptr as usize + size_of::<i32>(),
-            &mut output_len_bytes,
-        )?;
-        let output_ptr = i32::from_le_bytes(output_ptr_bytes);
-        let output_len = i32::from_le_bytes(output_len_bytes);
-
-        let mut response_bytes = Vec::from_iter(iter::repeat_n(0u8, output_len as usize));
-        memory.read(&store, output_ptr as usize, &mut response_bytes)?;
-        let response = O::decode(&mut &response_bytes[..])?;
-
-        Ok(response)
-    }
-
-    async fn call_wasm_func_without_return<I: Encode>(
-        func_name: &'static str,
-        module: Module,
-        linker: Arc<Linker<Connection>>,
-        db: libsql::Connection,
-        input: I,
-    ) -> anyhow::Result<()> {
-        let mut store = Store::new(&ENGINE, db);
-        let instance = linker.instantiate_async(&mut store, &module).await?;
-        let Some(memory) = instance.get_memory(&mut store, "memory") else {
-            anyhow::bail!("Could not load wasm memory.");
-        };
-        let malloc = instance
-            .get_typed_func::<(i32, i32), i32>(&mut store, "malloc")
-            .context(InvalidFunctionTypeCtx("malloc"))?;
-        let func = instance
-            .get_typed_func::<(i32, i32), ()>(&mut store, func_name)
-            .context(InvalidFunctionTypeCtx(func_name))?;
-
-        // Allocate the input in the WASM module
-        let input_bytes = input.encode();
-        let input_len = input_bytes.len() as i32;
-        let input_ptr = malloc.call_async(&mut store, (input_len, 1)).await?;
-        memory.write(&mut store, input_ptr as usize, &input_bytes)?;
-
-        // Run the named function from the WASM module
-        func.call_async(&mut store, (input_ptr, input_len))
-            .await
-            .inspect_err(|e| tracing::error!("{e}"))
-            .context(ErrorCallingFunctionCtx(func_name))?;
-
-        Ok(())
-    }
-}
-
-// FIXME: We need to apply fuel limits to WASM module execution, and we also need to watch memory
-// usage.
-impl LeafModule for LeafWasmModule {
-    fn id(&self) -> blake3::Hash {
-        self.id
-    }
-
-    fn filter_inbound(
-        &self,
-        input: IncomingEvent,
-        db: libsql::Connection,
-    ) -> Pin<Box<dyn Future<Output = anyhow::Result<Inbound>> + Send>> {
-        let module = self.module.clone();
-        let linker = self.linker.clone();
-        Box::pin(Self::call_wasm_func(
-            "filter_inbound",
-            module,
-            linker,
-            db,
-            input,
-        ))
-    }
-
-    fn filter_outbound(
-        &self,
-        input: EventRequest,
-        db: libsql::Connection,
-    ) -> Pin<Box<dyn Future<Output = anyhow::Result<Outbound>> + Send>> {
-        let module = self.module.clone();
-        let linker = self.linker.clone();
-
-        Box::pin(Self::call_wasm_func(
-            "filter_outbound",
-            module,
-            linker,
-            db,
-            input,
-        ))
-    }
-
-    fn process_event(
-        &self,
-        input: IncomingEvent,
-        db: libsql::Connection,
-    ) -> Pin<Box<dyn Future<Output = anyhow::Result<Process>> + Send>> {
-        let module = self.module.clone();
-        let linker = self.linker.clone();
-
-        Box::pin(Self::call_wasm_func(
-            "process_event",
-            module,
-            linker,
-            db,
-            input,
-        ))
-    }
-
-    fn fetch(
-        &self,
-        input: leaf_stream_types::FetchInput,
-        db: libsql::Connection,
-    ) -> Pin<Box<dyn Future<Output = anyhow::Result<()>> + Send>> {
-        let module = self.module.clone();
-        let linker = self.linker.clone();
-
-        Box::pin(Self::call_wasm_func_without_return(
-            "fetch", module, linker, db, input,
-        ))
-    }
-
-    fn init_db(
-        &self,
-        creator: String,
-        params: Vec<u8>,
-        db: libsql::Connection,
-    ) -> Pin<Box<dyn Future<Output = anyhow::Result<()>> + Send>> {
-        let module = self.module.clone();
-        let linker = self.linker.clone();
-
-        Box::pin(Self::call_wasm_func_without_return(
-            "init_db",
-            module,
-            linker,
-            db,
-            ModuleInit { creator, params },
-        ))
-    }
-}
-
-fn to_libsql(v: SqlValue) -> libsql::Value {
-    match v {
-        SqlValue::Null => libsql::Value::Null,
-        SqlValue::Integer(x) => libsql::Value::Integer(x),
-        SqlValue::Real(x) => libsql::Value::Real(x),
-        SqlValue::Text(x) => libsql::Value::Text(x),
-        SqlValue::Blob(x) => libsql::Value::Blob(x),
-    }
-}
-
-fn from_libsql(v: libsql::Value) -> SqlValue {
-    match v {
-        libsql::Value::Null => SqlValue::Null,
-        libsql::Value::Integer(x) => SqlValue::Integer(x),
-        libsql::Value::Real(x) => SqlValue::Real(x),
-        libsql::Value::Text(x) => SqlValue::Text(x),
-        libsql::Value::Blob(x) => SqlValue::Blob(x),
-    }
-}
